// src/types/mock-data.type.ts
export type MockData = {
  names: string[];
  descriptions: string[];
  postDates: string[];
  genres: string[];
  releases: string[];
  ratings: number[];
  previewVideos: string[];
  videos: string[];
  actors: string[][];
  directors: string[];
  durations: number[];
  commentsCounts: number[];
  userNames: string[];
  userEmails: string[];
  userAvatars: string[];
  userPasswords: string[];
  posters: string[];
  backgrounds: string[];
  backgroundColors: string[];
};